requirements.txt
streamlit==1.23.1
google-generativeai==0.7.2
numpy==1.24.3
pandas==2.0.2

